django_fcm_project/requirements.txt
asgiref==3.4.1
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coreapi==2.3.1
coreschema==0.0.4
Django==3.2.7
django-cors-headers==3.6.0
django-sslserver==0.22
djangorestframework==3.12.2
fcm-django==1.0.5
firebase==3.0.1
firebase-admin==5.0.2
gcloud==0.17.0
google-api-core==1.31.2
google-api-python-client==2.20.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud-core==2.0.0
google-cloud-firestore==2.3.1
google-cloud-storage==1.42.1
google-crc32c==1.1.2
google-resumable-media==2.0.2
googleapis-common-protos==1.53.0
grpcio==1.40.0
gunicorn==20.1.0
httplib2==0.19.1
idna==3.2
insultgenerator==1.0.3
itypes==1.2.0
Jinja2==3.0.1
jws==0.1.3
loremipsum==1.0.5
MarkupSafe==2.0.1
msgpack==1.0.2
oauth2client==3.0.0
packaging==21.0
proto-plus==1.19.0
protobuf==3.17.3
psycopg2-binary==2.9.1
pusher-push-notifications==2.0.1
py-notifier==0.3.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.4.3
PyJWT==2.1.0
pyparsing==2.4.7
python-jwt==2.0.1
pytz==2021.1
requests==2.26.0
requests-toolbelt==0.7.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
White-Noise==0.1.0
whitenoise==5.3.0
